# Reduce heap allocations when encoding