--- a/src/Base16.cs
+++ b/src/Base16.cs
@@ -213,7 +213,8 @@ public sealed class Base16(Base16Alphabet alphabet) : IBaseCoder, IBaseStreamCod
             return string.Empty;
         }
 
-        Span<char> output = new char[GetSafeCharCountForEncoding(bytes)];
+        int outputLen = GetSafeCharCountForEncoding(bytes);
+        Span<char> output = outputLen < 1024 ? stackalloc char[outputLen] : new char[outputLen];
         internalEncode(bytes, output, Alphabet.Value);
         return new string(output);
     }