Reduce heap allocations, that needs to be garbage collected, when encoding.
Using [stack allocations](https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/operators/stackalloc) instead, when it makes sense.

I have readded `MemoryDiagnoser` (no longer default) to encoder benchmarks to see reduced allocations & GC, besides wallclock performance.

**Before:**

| Method                                 | Mean      | Error    | StdDev    | Median    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------- |----------:|---------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| DotNet_Base64                          |  39.58 ns | 0.824 ns |  0.980 ns |  39.29 ns |  1.00 |    0.03 | 0.0318 |     200 B |        1.00 |
| SimpleBase_Base16_UpperCase            | 120.72 ns | 1.808 ns |  1.510 ns | 120.98 ns |  3.05 |    0.08 | 0.0892 |     560 B |        2.80 |
| SimpleBase_Base32_CrockfordWithPadding | 202.62 ns | 4.096 ns | 10.498 ns | 208.17 ns |  5.12 |    0.29 | 0.0739 |     464 B |        2.32 |
| SimpleBase_Base85_Z85                  | 230.25 ns | 4.599 ns |  9.078 ns | 231.50 ns |  5.82 |    0.27 | 0.0892 |     560 B |        2.80 |
| SimpleBase_Base58_Bitcoin              |  62.06 ns | 1.289 ns |  1.266 ns |  62.06 ns |  1.57 |    0.05 | 0.0497 |     312 B |        1.56 |

**With stack allocations:**

| Method                                 | Mean      | Error    | StdDev    | Median    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------- |----------:|---------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| DotNet_Base64                          |  39.43 ns | 0.587 ns |  0.549 ns |  39.44 ns |  1.00 |    0.02 | 0.0318 |     200 B |        1.00 |
| SimpleBase_Base16_UpperCase            | 114.88 ns | 1.392 ns |  1.234 ns | 114.57 ns |  2.91 |    0.05 | 0.0446 |     280 B |        1.40 |
| SimpleBase_Base32_CrockfordWithPadding | 191.03 ns | 3.880 ns | 10.221 ns | 195.69 ns |  4.85 |    0.27 | 0.0370 |     232 B |        1.16 |
| SimpleBase_Base85_Z85                  | 199.14 ns | 4.072 ns | 11.944 ns | 199.54 ns |  5.05 |    0.31 | 0.0293 |     184 B |        0.92 |
| SimpleBase_Base58_Bitcoin              |  54.27 ns | 1.161 ns |  2.691 ns |  53.67 ns |  1.38 |    0.07 | 0.0242 |     152 B |        0.76 |

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3775)  
Intel Core i7-10750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores  
.NET SDK 8.0.408  
  [Host]     : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2  
  DefaultJob : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2